refactor(dtos): move LoginResponseDTO and UserResponseDTO classes to a new response package within the dto package (#11)
